+++ test/faker/currency_test.exs
@@ -1,0 +1,5 @@
+defmodule Faker.CurrencyTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Currency
+end