+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 ### Added
 
 - `Faker.Gov.Us.ein/0` [[@artkay](https://github.com/artkay)]
+- `Faker.Currency.code/0` and `Faker.Currency.symbol/0` [[@brunohkbx](https://github.com/brunohkbx)]
 
 ### Changed